Initial tweaks to tracker icons animation
